# Modify XmlSerializators to use typed XmlWriter.WriteValue method