Modify XmlSerializators to use typed XmlWriter.WriteValue method instead of converting to string when value is primitive.
Once [Optimization](https://github.com/dotnet/runtime/pull/75411) goes in, this should reduce allocations.

Here are some benchmarks from performance repo:
BenchmarkDotNet=v0.13.1.1847-nightly, OS=Windows 10 (10.0.19043.1889/21H1/May2021Update)
11th Gen Intel Core i9-11900K 3.50GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK=7.0.100-preview.7.22377.5
  [Host]     : .NET 7.0.0 (7.0.22.37506), X64 RyuJIT AVX2
  Job-LGCOMF : .NET 8.0.0 (42.42.42.42424), X64 RyuJIT AVX2

PowerPlanMode=00000000-0000-0000-0000-000000000000  Toolchain=CoreRun  IterationTime=250.0000 ms  
MaxIterationCount=20  MinIterationCount=15  WarmupCount=1  

Benchmark : MicroBenchmarks.Serializers.Xml_FromStream_IndexViewModel
Before:
|                 Method |     Mean |    Error |   StdDev |   Median |      Min |      Max |  Gen 0 |  Gen 1 | Allocated |
|----------------------- |---------:|---------:|---------:|---------:|---------:|---------:|-------:|-------:|----------:|
| DataContractSerializer | 85.41 μs | 0.670 μs | 0.594 μs | 85.53 μs | 83.64 μs | 86.16 μs | 4.4521 | 0.6849 |  37.89 KB |
|          XmlSerializer | 48.82 μs | 0.372 μs | 0.348 μs | 48.71 μs | 48.45 μs | 49.51 μs | 5.2376 | 0.7759 |  44.24 KB |

After both PRs are in a branch:
|                 Method |     Mean |    Error |   StdDev |   Median |      Min |      Max |  Gen 0 |  Gen 1 | Allocated |
|----------------------- |---------:|---------:|---------:|---------:|---------:|---------:|-------:|-------:|----------:|
| DataContractSerializer | 85.60 μs | 0.227 μs | 0.190 μs | 85.56 μs | 85.39 μs | 86.07 μs | 4.4353 | 0.6824 |  37.89 KB |
|          XmlSerializer | 47.76 μs | 0.900 μs | 0.925 μs | 47.80 μs | 45.41 μs | 49.34 μs | 5.2765 | 0.7817 |  43.63 KB |