@TrayanZapryanov 2022-09-12
Looks quote promising as results once both PRs are applied:
Benchmark :MicroBenchmarks.Serializers.Xml_ToStream_MyEventsListerViewModel
Before:
|                 Method |     Mean |   Error |  StdDev |   Median |      Min |      Max |   Gen 0 | Allocated |
|----------------------- |---------:|--------:|--------:|---------:|---------:|---------:|--------:|----------:|
| DataContractSerializer | 401.8 μs | 5.09 μs | 4.25 μs | 403.0 μs | 389.9 μs | 405.9 μs | 11.3086 | 102.98 KB |
|          XmlSerializer | 251.3 μs | 3.05 μs | 2.55 μs | 251.4 μs | 243.6 μs | 253.9 μs |  8.0645 |  69.06 KB |

After:
|                 Method |     Mean |   Error |  StdDev |   Median |      Min |      Max |   Gen 0 | Allocated |
|----------------------- |---------:|--------:|--------:|---------:|---------:|---------:|--------:|----------:|
| DataContractSerializer | 415.2 μs | 4.52 μs | 4.23 μs | 415.8 μs | 401.3 μs | 419.2 μs | 11.6667 | 102.98 KB |
|          XmlSerializer | 250.9 μs | 3.04 μs | 2.54 μs | 251.9 μs | 245.0 μs | 252.8 μs |  4.0323 |  34.28 KB |